Refactored ReadabiltyCreator

+++ Calculator/Calculator.Logic/Parsing/ConversionTokenizer/ConversionFactors.cs
@@ -3,18 +3,18 @@ namespace Calculator.Logic.Parsing.ConversionTokenizer
     public static class ConversionFactors
     {
         //Metric
-        public const decimal MetricDivisionOneMillion = (decimal) 1.0E-6;
-        public const decimal MetricDivisionOneThousand = (decimal) 1.0E-3;
-        public const decimal MetricDivisionTenThousand = (decimal) 1.0E-4;
-        public const decimal MetricDivisionOneHundred = (decimal) 1.0E-2;
+        public const decimal MetricDivisionOneMillion = 1.0E-6M;
+        public const decimal MetricDivisionOneThousand = 1.0E-3M;
+        public const decimal MetricDivisionTenThousand = 1.0E-4M;
+        public const decimal MetricDivisionOneHundred = 1.0E-2M;
         public const decimal MultiplicationByOne = 1;
-        public const decimal MetricMultiplicationOneHundred = (decimal) 1.0E2;
-        public const decimal MetricMultiplicationOneThousand = (decimal) 1.0E3;
-        public const decimal MetricMultiplicationOneMillion = (decimal) 1.0E6;
-        public const decimal MetricMultiplicationMeterToha = (decimal) 1.0E4;
+        public const decimal MetricMultiplicationOneHundred = 1.0E2M;
+        public const decimal MetricMultiplicationOneThousand = 1.0E3M;
+        public const decimal MetricMultiplicationOneMillion = 1.0E6M;
+        public const decimal MetricMultiplicationMeterToha = 1.0E4M;
         //Imperial Length
-        public const decimal ThouToFeet = (decimal) 1/12000;
-        public const decimal InchToFeet = (decimal) 1/12;
+        public const decimal ThouToFeet = 1M/12000;
+        public const decimal InchToFeet = 1M/12;
         public const decimal YardToFeet = 3;
         public const decimal ChainToFeet = 66;
         public const decimal FurlongToFeet = 660;
@@ -22,7 +22,7 @@ namespace Calculator.Logic.Parsing.ConversionTokenizer
         public const decimal LeagueToFeet = 15840;
         public const decimal FathomToFeet = 608001;
         //Imperial Area
-        public const decimal PerchToSquareFoot = (decimal) 272.25;
+        public const decimal PerchToSquareFoot = 272.25M;
         public const decimal RoodToSquareFoot = 10890;
         public const decimal AcreToSquareFoot = 43560;
         //Imperial Volume
@@ -31,9 +31,9 @@ namespace Calculator.Logic.Parsing.ConversionTokenizer
         public const decimal QuartToFluidOunce = 40;
         public const decimal GallonToFluidOunce = 160;
         //Imperial Mass
-        public const decimal GrainToPound = (decimal) 1/7000;
-        public const decimal DrachmToPound = (decimal) 1/256;
-        public const decimal OunceToPound = (decimal) 1/16;
+        public const decimal GrainToPound = 1M/7000;
+        public const decimal DrachmToPound = 1M/256;
+        public const decimal OunceToPound = 1M/16;
         public const decimal StoneToPound = 14;
         public const decimal HundredWeightToPound = 112;
         public const decimal ImperialTonToPound = 2240;
